Fixed some visual effects
* fixed tails of some effects
* activated randomization based on fps
* fixed full set effects by adding randomizing

+++ Source Main 5.2/source/_struct.h
@@ -524,7 +524,7 @@ typedef struct
     OBJECT* Target;
     vec3_t      TargetPosition;
     BYTE        byOnlyOneRender;
-    int			NumTails;
+    float		NumTails;
     int			MaxTails;
     vec3_t		Tails[MAX_TAILS][4];
     float  		LifeTime;
